fix: codejar should not be an esmodule in a browser environment
